JobStateMutator subscribes to JobCancelledEvent
